requirements.txt
numpy==1.21.1
boto3==1.18.21
python-dotenv==0.19.0
